+++ src/common/database/queries/select-query-text.js
@@ -31,16 +31,6 @@ function writeDocumentFileSelect(vOffset, vSize)
 }
 
 
-// Select individual advertisement case using it's code key.
-function writeCaseKeySelect(vKey)
-{
-	var queryFormat = "SELECT caseCode FROM CaseFile WHERE caseCode = ?";
-	var queryParameters = [vKey];
-	var writeRes = mysql.format(queryFormat, queryParameters);
-	return writeRes;
-}
-
-
 // Select most recent determination date from advertisement cases.
 function writeLatestDeterminationDateSelect()
 {
@@ -53,56 +43,11 @@ function writeLatestDeterminationDateSelect()
 	return writeRes;
 }
 
-// Select group of rows from a table.
-function writeExportRowSelect(vTableName, vSortKey, vOffset, vSize)
-{
-	var queryFormat = "";
-	var queryParameters = [vTableName, vSortKey, vOffset, vSize];
-	
-	queryFormat += "SELECT * FROM ?? ";
-	queryFormat += "ORDER BY ??";
-	queryFormat += "LIMIT ?,?";
-	
-	var writeRes = mysql.format(queryFormat, queryParameters);
-	return writeRes;
-}
-
-
-// Select advertisement case entries with blank values.
-function writeUnknownEntriesSelect(vOffset, vSize)
-{
-	var queryFormat = "";
-	var queryParameters = [vOffset, vSize];
-	
-	queryFormat += "SELECT ";
-	queryFormat += "caseEntryID, caseCode, advertiserID, descriptionText, ";
-	queryFormat += "determinationFlag, productCategoryID, mediaTypeID, determinationDate ";
-	queryFormat += "FROM CaseFile WHERE (";
-	
-	queryFormat += "(advertiserID IS NULL) OR ";
-	queryFormat += "(determinationFlag IS NULL) OR ";
-	queryFormat += "(productCategoryID IS NULL) OR ";
-	queryFormat += "(mediaTypeID IS NULL) OR ";
-	queryFormat += "(determinationDate IS NULL)";
-	
-	queryFormat += ") AND ";
-	queryFormat += "(documentFileURL IS NOT NULL) ";
-	
-	queryFormat += "ORDER BY caseEntryID ";
-	queryFormat += "LIMIT ?,?";
-	
-	var writeRes = mysql.format(queryFormat, queryParameters);
-	return writeRes;
-}
-
 
 
 module.exports =
 {
 	writeListKey: writeListKeySelect,
 	writeDocumentFile: writeDocumentFileSelect,
-	writeCaseKey: writeCaseKeySelect,
-	writeLatestDeterminationDate: writeLatestDeterminationDateSelect,
-	writeExportRow: writeExportRowSelect,
-	writeUnknownEntries: writeUnknownEntriesSelect
+	writeLatestDeterminationDate: writeLatestDeterminationDateSelect
 };