+++ src/common/entry/entry-error-text.js
@@ -66,72 +66,12 @@ function writeStringAlreadyTakenText(vDesc, vTakenItem)
 }
 
 
-// Number already taken.
-function writeNumberAlreadyTakenText(vDesc, vTakenNumber)
-{
-	var writeRes = "";
-	
-	writeRes += "The ";
-	writeRes += vDesc;
-	writeRes += " number ";
-	writeRes += vTakenNumber;
-	writeRes += " is already taken.";
-	
-	return writeRes;
-}
-
-// Unknown string.
-function writeStringDoesNotExistText(vDesc, vUnknownString)
-{
-	var writeRes = "";
-	
-	writeRes += "The ";
-	writeRes += vDesc;
-	writeRes += " item '";
-	writeRes += vUnknownString;
-	writeRes += "' does not exist.";
-	
-	return writeRes;
-}
-
-
 // Page number parameter missing.
 function writePageMissingText(vDesc)
 {
 	var writeRes = "Page number parameter is missing from " + vDesc;
 	return writeRes;
 }
-
-// Invalid number key.
-function writeInvalidNumberKeyText(vDesc)
-{
-	var writeRes = vDesc + " must be a valid positive whole number.";
-	return writeRes;
-}
-
-
-// Invalid case code format.
-function writeInvalidCaseCodeText(vCode)
-{
-	var writeRes = "";
-	
-	writeRes += "The string '";
-	writeRes += vCode;
-	writeRes += "' is not a valid case code. Must follow the format: ";
-	writeRes += "'dddd-yy'";
-	
-	return writeRes;
-}
-
-// Invalid determination date.
-function writeInvalidDeterminationDateText()
-{
-	var writeRes = "Invalid determination date. Must use the format 'yyyy-mm-dd'";
-	return writeRes;
-}
-
-
-
 
 
 module.exports =
@@ -141,10 +81,5 @@ module.exports =
 	writeNumberRange: writeNumberRangeText,
 	writeEmpty: writeEmptyText,
 	writeStringAlreadyTaken: writeStringAlreadyTakenText,
-	writeNumberAlreadyTaken: writeNumberAlreadyTakenText,
-	writeStringDoesNotExist: writeStringDoesNotExistText,
-	writePageMissing: writePageMissingText,
-	writeInvalidNumberKey: writeInvalidNumberKeyText,
-	writeInvalidCaseCode: writeInvalidCaseCodeText,
-	writeInvalidDeterminationDate: writeInvalidDeterminationDateText
+	writePageMissing: writePageMissingText
 };