+++ src/common/interface/general/data-desc.js
@@ -1,15 +1,1 @@
-// Import and Export data file description text.
-
-const advertiserDesc = "Advertiser Data File";
-const productCategoryDesc = "Product Category Data File";
-const mediaTypeDesc = "Media Type Data File";
-const caseFileDesc = "Advertisement Cases Data File";
-
-
-module.exports =
-{
-	advertiser: advertiserDesc,
-	productCategory: productCategoryDesc,
-	mediaType: mediaTypeDesc,
-	caseFile: caseFileDesc
-};
+// EMPTY