+++ stored-paths.js
@@ -5,9 +5,7 @@ const path = require("path");
 const databaseFolderPath = path.join(".", "src", "common", "database", "scripts");					// Database scripts folder
 const pageFolderPath = path.join(".", "pages");														// HTML pages folder.
 const reportsFolderPath = path.join(".", "reports");												// Advertisement case reports root folder.
-const exportFolderPath = path.join(".", "export");													// Default export folder.
 const testDownloadFolderPath = path.join(".", "test-download");										// Test download folder.
-const unknownOutputFilePath = path.join(".", "unknown-entries.csv");								// Default unknown entries file.
 
 const logFilePath = path.join(".", "log.txt");														// Log file path.
 const dbCredsPath = path.join(".", "db-creds.json");												// Database credentials file path.
@@ -21,9 +19,7 @@ module.exports =
 {
 	pageFolder: pageFolderPath,
 	reportsFolder: reportsFolderPath,
-	exportFolder: exportFolderPath,
 	testDownloadFolder: testDownloadFolderPath,
-	unknownOutputFile: unknownOutputFilePath,
 	logFile: logFilePath,
 	dbCreds: dbCredsPath,
 	dbDefinition: dbDefinitionPath,